backend/requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.27
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
pydantic[email]>=2.6
pydantic-settings>=2.1
python-dotenv>=1.0
python-multipart>=0.0.9
passlib[bcrypt]>=1.7
httpx>=0.26
openai>=1.12
